Update graphsync to 2.0 branch - absolute minimum (#300)
* feat(deps): update graphsync to 2.0 branch

Provide the absolute minimum required to work with the latest version of
graphsync with the 2.0 protocl.

This does use a very unfortunate hack to deal with the change to
extension type -- it now deserializes/serializes from basicnode.Any
in order to get back to a []byte type

* style(mod): tidy

* refactor(message): create To/From IPLD funcs

* fix(impl): add one more int conversion

* feat(msg): ipld-prime messages, compat testing w/ cbor-gen

* fix(msg): replace message 1_1 with 1_1prime form

* chore(msg): rename struct fields, use renames for consistent encoding

* fix(msg): remove through-bytes roundtrip for conversions

* fix(msg): fixed upstream bindnode bug, added some temporary debugging

* fix(graphsync): compare Nodes with ipld.DeepEqual

* fix(deps): update ipld-prime with bindnode tuple fix

* fix(msg): remove debug prints, fix imports

* fix(cbor-gen): make go generate work again in old msgs

* chore(deps): update go-ipld-prime to new master with tuple fix

* feat(graphsync): remove peer id tracking

no longer need to utilize peer id to tract graphsync requests since RequestIDs are unique

* feat(deps): update to tagged go-graphsync

+++ tracing/tracing_test.go
@@ -15,8 +15,8 @@ import (
 func TestSpansIndex(t *testing.T) {
 	ctx := context.Background()
 	peers := testutil.GeneratePeers(2)
-	chid1 := datatransfer.ChannelID{Initiator: peers[0], Responder: peers[1], ID: datatransfer.TransferID(rand.Uint64())}
-	chid2 := datatransfer.ChannelID{Initiator: peers[0], Responder: peers[1], ID: datatransfer.TransferID(rand.Uint64())}
+	chid1 := datatransfer.ChannelID{Initiator: peers[0], Responder: peers[1], ID: datatransfer.TransferID(rand.Uint32())}
+	chid2 := datatransfer.ChannelID{Initiator: peers[0], Responder: peers[1], ID: datatransfer.TransferID(rand.Uint32())}
 	testCases := map[string]struct {
 		operation      func(ctx context.Context, si *tracing.SpansIndex)
 		expectedTraces []string
